fix(sm90): cast activation probe attributes
Why:
CUDA 13 reports shared and local resource sizes as size_t, which cannot be
list-initialized into the probe int64 result without an explicit cast.

What:
Cast all cudaFuncAttributes and occupancy values to int64_t.

Validation:
git diff --check
Remote rebuild pending on GPU9.

+++ tests/terminal_activation_probe.cu
@@ -168,9 +168,13 @@ std::vector<int64_t> attributes() {
                     &occupancy, cluster_kernel, &config) == cudaSuccess,
                 "cluster occupancy query failed");
     return {
-        reference.numRegs, reference.sharedSizeBytes, reference.localSizeBytes,
-        clustered.numRegs, clustered.sharedSizeBytes, clustered.localSizeBytes,
-        occupancy,
+        static_cast<int64_t>(reference.numRegs),
+        static_cast<int64_t>(reference.sharedSizeBytes),
+        static_cast<int64_t>(reference.localSizeBytes),
+        static_cast<int64_t>(clustered.numRegs),
+        static_cast<int64_t>(clustered.sharedSizeBytes),
+        static_cast<int64_t>(clustered.localSizeBytes),
+        static_cast<int64_t>(occupancy),
     };
 }
